# SubastasWeb/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Subastas",
        "description": "Documentación de la API de Subastas",
        "contact": {
            "email": "contacto@subastas.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/admin/usuarios-pendientes": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Listar casas de remate pendientes de aprobación (paginado)",
                "operationId": "6640c0779752fc9ef54b71752d2f0e88",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de casas pendientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "idFiscal": {
                                                        "type": "string",
                                                        "example": "212345670018"
                                                    },
                                                    "aprobado_en": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "aprobado_por": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "usuario": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Casa Aurora"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "casa@example.com"
                                                            },
                                                            "telefono": {
                                                                "type": "string",
                                                                "example": "099123456"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/casas-activas": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Listar casas de remate activas (paginado)",
                "operationId": "72dcef9b203a64839f8c450e025f71f9",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de casas activas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "idFiscal": {
                                                        "type": "string",
                                                        "example": "211234560019"
                                                    },
                                                    "aprobado_en": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-06-19T15:45:00Z"
                                                    },
                                                    "aprobado_por": {
                                                        "type": "string",
                                                        "example": "admin@example.com"
                                                    },
                                                    "usuario": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 6
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Casa Solis"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "solis@example.com"
                                                            },
                                                            "telefono": {
                                                                "type": "string",
                                                                "example": "098765432"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/aprobar-casa/{usuarioId}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Aprobar una casa de remate",
                "operationId": "1a39b38234d52b1d2486505f24a7fc25",
                "parameters": [
                    {
                        "name": "usuarioId",
                        "in": "path",
                        "description": "ID del usuario asociado a la casa de remate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Casa aprobada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Casa de remate aprobada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Casa no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Casa de remate no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/desaprobar-casa/{usuarioId}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Desaprobar o desactivar una casa de remate",
                "operationId": "b3fb78369507a14e555fc5fa8b4055a7",
                "parameters": [
                    {
                        "name": "usuarioId",
                        "in": "path",
                        "description": "ID del usuario asociado a la casa de remate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Casa desaprobada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Casa de remate desaprobada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Casa no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Casa de remate no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/eliminar-usuario/{usuario_id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Eliminar un usuario y su perfil asociado",
                "operationId": "5260171b6697649d9623d7469ac48749",
                "parameters": [
                    {
                        "name": "usuario_id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario y perfil asociado eliminados con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/usuarios": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Listar usuarios por rol (paginado)",
                "operationId": "313ca5c896442b45a40979113c7cb2b7",
                "parameters": [
                    {
                        "name": "rol",
                        "in": "query",
                        "description": "Rol de los usuarios a listar: cliente, rematador o casa_remate",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "cliente",
                                "rematador",
                                "casa_remate"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de resultados por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de usuarios por rol",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Rol inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/articulos": {
            "get": {
                "tags": [
                    "Articulos"
                ],
                "summary": "Obtener todos los articulos",
                "operationId": "021362b92b70d8c24f566bb0db7178d5",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Articulos"
                ],
                "summary": "Crear un nuevo artículo",
                "operationId": "1576d3467aeea24559ab2a70e9a963d9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "imagenes",
                                    "especificacion",
                                    "disponibilidad",
                                    "condicion",
                                    "vendedor_id"
                                ],
                                "properties": {
                                    "imagenes": {
                                        "type": "string"
                                    },
                                    "especificacion": {
                                        "type": "string"
                                    },
                                    "disponibilidad": {
                                        "type": "boolean"
                                    },
                                    "condicion": {
                                        "type": "string"
                                    },
                                    "vendedor_id": {
                                        "type": "integer"
                                    },
                                    "lote_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Artículo creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
...
        "/api/forgot-password": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Enviar enlace para restablecer contraseña",
                "operationId": "27192d4d04cb48a2f83216053554a56c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Enlace de restablecimiento enviado."
                    },
                    "422": {
                        "description": "Email no válido o no registrado."
                    },
                    "500": {
                        "description": "Error al enviar el enlace."
                    }
                }
            }
        },
...
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada con éxito."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ForgotPasswordRequest": {
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "elchati@adinet.com.uy"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}